Rerun fi (#151)
* Toggle to save options and results set to True by default

* enum for rerunFI

* function to load previous FI options

* black format

* Checks if previous FI experiments have been run, and if so, displays option to user to delete everything and re run. Added function to delete all the previous results if user selects so

* Added function to load previous json file with FI options. Added function to delete all FI generated directories

* Correction in the name used to save the fuzzy options, before the FI name was being used

* added function to find previous FI opt

* changed to pass flake8

* Move load_fi_options to services.configuration

* Find previous results moved to services.experiments. Function now checks if ANY FI importance result is saved, if so, returns True, if not, returns False

* Move delete_previous_FI_results to services.experiments. Now algorithm creates a list and iterates over it

* correction to delete directory function to be able to delete files as well (necessary for json files)

* Remove data choices. Correct imports of functions to check and delete previous FI results

* Create delete_file function

* fi_json_options converted to FeatureImportanceOptions class

* delete_previous_FI_results selectively uses delete_file or delete_directory depending on the nature of the path

* Correction of argstring

* Correction to return type, argstring and handles TypeError. No general except allowed by flake8

* previous_results_exist

+++ biofefi/utils/utils.py
@@ -117,7 +117,7 @@ def cancel_pipeline(p: Process):
 
 
 def delete_directory(path: Path):
-    """Delete the previous results from the log_dir
+    """Delete a directory and all of its contents.
 
     Returns:
         None
@@ -125,4 +125,18 @@ def delete_directory(path: Path):
     Args:
         path: The path to the log directory
     """
-    shutil.rmtree(path, ignore_errors=True)
+    if path.is_dir():
+        shutil.rmtree(path, ignore_errors=True)  # Remove the directory
+
+
+def delete_file(path: Path):
+    """Delete a file.
+
+    Returns:
+        None
+
+    Args:
+        path: The path to the log directory
+    """
+    if path.is_file():
+        path.unlink()  # Remove the file
